+++ stores/useRevCatStore.tsx
@@ -2,12 +2,12 @@ import { create } from "zustand";
 
 interface useRevCatStoreState {
   products: any;
-  setProducts: (products: { weekly: any; annual: any }) => void;
+  setProducts: (products: { weekly: any; annual: any, monthly: any }) => void;
 }
 
 const useRevCatStore = create<useRevCatStoreState>((set, get) => ({
   products: null,
-  setProducts: (products: { weekly: any; annual: any }) => set({ products }),
+    setProducts: (products: { weekly: any; annual: any, monthly: any }) => set({ products }),
 }));
 
 export default useRevCatStore;