+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2026-06-08
+
+### Added
+- Integrated tiktok-sdk
+
+### Changed
+- Subscription period switched from yearly to monthly
+- Updated the weekly subscription price
+
+### Fixed
+- Resolved an issue where users were not redirected to the paywall when restarting the app
+
 ## [1.0.7] - 2026-05-30
 
 ### Added